Research-Agent/pyproject.toml
# Generated by Copilot
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-agent"
version = "0.1.0"
description = "AI Research Agent"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai",
    "logfire",
    "aiohttp",
    "httpx",
    "tenacity",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]
